+++ CreateTables/CreateProduct.sql
@@ -1,6 +1,6 @@
 -- Create Product Table
 CREATE TABLE Product (
-    Product_ID INTEGER PRIMARY KEY,
+    Product_ID NUMBER(15,0) PRIMARY KEY,
     Product_Name VARCHAR2(100),
     Product_Price NUMBER(10,2)
 );
